Separate database packages
This is a very small refactor which aligns more with the spirit of our
repo which is that meaningful exports are placed in their own file and
consolidated by an index.ts.
